feat: support for disabling .html in links
