+++ src/components/App/useLangs.ts
@@ -1,14 +1,17 @@
 import type {AppProps} from './index';
+import type {Lang} from '@diplodoc/components';
 
-import {Lang, getLangPath} from '@diplodoc/components';
+import {getLangPath} from '@diplodoc/components';
 import {useCallback, useMemo} from 'react';
 
-export function useLangs({router, lang, langs}: AppProps) {
+export function useLangs({router, lang, langs}: AppProps, skipHtmlExtension?: boolean) {
     const onChangeLang = useCallback(
         (newLang: `${Lang}` | Lang) => {
-            window.location.replace(getLangPath(router, newLang, window.location.href));
+            window.location.replace(
+                getLangPath(router, newLang, window.location.href, skipHtmlExtension),
+            );
         },
-        [router],
+        [router, skipHtmlExtension],
     );
 
     return useMemo(
