+++ src/interceptors/leading-page-links.js
@@ -25,36 +25,4 @@
             }
         })(Element.prototype);
     }
-
-    if (typeof document !== 'undefined') {
-        document.addEventListener('click', (event) => {
-            const href = event.target.href;
-            const locationOrigin = window.location.origin;
-
-            if (
-                event.target.matches('.dc-doc-layout__center a') &&
-                href.startsWith(locationOrigin)
-            ) {
-                event.preventDefault();
-
-                const mainFileName = 'index';
-                const extention = '.html';
-
-                if (href.endsWith('/')) {
-                    window.location.href = `${href}${mainFileName}${extention}`;
-                    return;
-                }
-
-                // https://../file-name, https://../file-name#fragment
-                const splitedHref = href.split('#');
-                if (splitedHref.length > 0 && !splitedHref[0].endsWith(extention)) {
-                    splitedHref[0] += extention;
-                    window.location.href = splitedHref.join('#');
-                    return;
-                }
-
-                window.location.href = href;
-            }
-        });
-    }
 })();
